# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Career OpenClaw API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/app/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "description": "List job records discovered in the current workspace.",
        "operationId": "list_jobs_api_app_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by job status: discovered|reportable|invalid|stale",
              "title": "Status"
            },
            "description": "Filter by job status: discovered|reportable|invalid|stale"
          },
          {
            "name": "include_report_summary",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Join latest active job report for role category/seniority/confidence fields",
              "default": false,
              "title": "Include Report Summary"
            },
            "description": "Join latest active job report for role category/seniority/confidence fields"
          },
          {
            "name": "favorites_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only return jobs bookmarked in this workspace",
              "default": false,
              "title": "Favorites Only"
            },
            "description": "Only return jobs bookmarked in this workspace"
          },
          {
            "name": "not_interested_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only return jobs dismissed as not interested in this workspace",
              "default": false,
              "title": "Not Interested Only"
            },
            "description": "Only return jobs dismissed as not interested in this workspace"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/app/jobs/import": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Import Job",
        "description": "Import a single job by URL: fetch JD, extract fields, persist.",
        "operationId": "import_job_api_app_jobs_import_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/app/jobs/batch-archive": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Batch Archive",
        "description": "Archive multiple jobs at once.",
        "operationId": "batch_archive_api_app_jobs_batch_archive_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchArchiveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchArchiveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/app/jobs/batch-analyze": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Batch Analyze",
        "description": "Trigger fit analysis for multiple jobs.\n\nJobs with an existing job report get a fit_report run directly.\nJobs without a job report get a job_report run first, which\nauto-chains into a fit_report run upon completion.",
        "operationId": "batch_analyze_api_app_jobs_batch_analyze_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchAnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchAnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchAnalyzeRequest": {
        "properties": {
          "job_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Job Ids"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          }
        },
        "type": "object",
        "required": [
          "job_ids"
        ],
        "title": "BatchAnalyzeRequest"
      },
      "BatchAnalyzeResponse": {
        "properties": {
          "run_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Run Ids"
          },
          "skipped": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skipped"
          },
          "report_first": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Report First",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "run_ids",
          "skipped"
        ],
        "title": "BatchAnalyzeResponse"
      },
      "BatchArchiveRequest": {
        "properties": {
          "job_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Job Ids"
          }
        },
        "type": "object",
        "required": [
          "job_ids"
        ],
        "title": "BatchArchiveRequest"
      },
      "BatchArchiveResponse": {
        "properties": {
          "archived_count": {
            "type": "integer",
            "title": "Archived Count"
          }
        },
        "type": "object",
        "required": [
          "archived_count"
        ],
        "title": "BatchArchiveResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobImportRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "JobImportRequest",
        "description": "Import a single job by URL.",
        "examples": [
          {
            "url": "https://boards.greenhouse.io/acme/jobs/123"
          }
        ]
      },
      "JobImportResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobRead"
          },
          "created": {
            "type": "boolean",
            "title": "Created"
          },
          "jd_fetched": {
            "type": "boolean",
            "title": "Jd Fetched"
          }
        },
        "type": "object",
        "required": [
          "job",
          "created",
          "jd_fetched"
        ],
        "title": "JobImportResponse",
        "description": "Result of a job import."
      },
      "JobList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "JobList"
      },
...
    }
  }
}